59: feat: update openapi spec and server to include publish and setreplica r=tiagolobocastro a=tiagolobocastro

feat: update openapi spec and autogenerated code
Explicitly specify min/max for some of the integer types we use.

feat: update openapispec and model to include publish and setreplica
Add new openapi operations for creating a volume target (this means we create a nexus)
and new operation to change a volume's replica count.

feat: update rest server to include setreplica and publish
Refactor publish to return the volume.

+++ common/src/types/v0/message_bus/child.rs
@@ -13,7 +13,7 @@ pub struct Child {
     /// state of the child
     pub state: ChildState,
     /// current rebuild progress (%)
-    pub rebuild_progress: Option<i32>,
+    pub rebuild_progress: Option<u8>,
 }
 
 impl From<Child> for models::Child {
