+++ openapi/src/models/volume_spec.rs
@@ -24,17 +24,17 @@ pub struct VolumeSpec {
     pub labels: Vec<String>,
     /// Number of front-end paths.
     #[serde(rename = "num_paths")]
-    pub num_paths: i32,
+    pub num_paths: u8,
     /// Number of children the volume should have.
     #[serde(rename = "num_replicas")]
-    pub num_replicas: i32,
+    pub num_replicas: u8,
     #[serde(rename = "operation", skip_serializing_if = "Option::is_none")]
     pub operation: Option<crate::models::VolumeSpecOperation>,
     #[serde(rename = "protocol")]
     pub protocol: crate::models::Protocol,
     /// Size that the volume should be.
     #[serde(rename = "size")]
-    pub size: i64,
+    pub size: u64,
     #[serde(rename = "state")]
     pub state: crate::models::SpecState,
     /// The node where front-end IO will be sent to
@@ -49,10 +49,10 @@ impl VolumeSpec {
     /// VolumeSpec using only the required fields
     pub fn new(
         labels: impl IntoVec<String>,
-        num_paths: impl Into<i32>,
-        num_replicas: impl Into<i32>,
+        num_paths: impl Into<u8>,
+        num_replicas: impl Into<u8>,
         protocol: impl Into<crate::models::Protocol>,
-        size: impl Into<i64>,
+        size: impl Into<u64>,
         state: impl Into<crate::models::SpecState>,
         uuid: impl Into<uuid::Uuid>,
     ) -> VolumeSpec {
@@ -71,11 +71,11 @@ impl VolumeSpec {
     /// VolumeSpec using all fields
     pub fn new_all(
         labels: impl IntoVec<String>,
-        num_paths: impl Into<i32>,
-        num_replicas: impl Into<i32>,
+        num_paths: impl Into<u8>,
+        num_replicas: impl Into<u8>,
         operation: impl Into<Option<crate::models::VolumeSpecOperation>>,
         protocol: impl Into<crate::models::Protocol>,
-        size: impl Into<i64>,
+        size: impl Into<u64>,
         state: impl Into<crate::models::SpecState>,
         target_node: impl Into<Option<String>>,
         uuid: impl Into<uuid::Uuid>,
